Bucher method for adjusted treatment effect

Given two estimated treatment effects of A vs. C and B vs. C (i.e. two point estimates and corresponding standard errors), derive the adjusted treatment of A vs. B using Bucher method, with two-sided confidence limits and Z-test p-value

Usage

bucher(trt, com, conf_lv = 0.95)

Arguments

trt

a list with two named scalars for the study with interested experimental arm, one named 'est' for the point estimate, and the other named 'se' for the standard error

com

same as trt, but for the study with interested control arm

conf_lv

a numerical scalar, prescribe confidence level to derive two-sided confidence interval for the adjusted treatment effect

Value

a list with 5 elements,

est

a scalar, point estimate of the adjusted treatment effect

se

a scalar, standard error of the adjusted treatment effect (i.e. est in return)

ci_l

a scalar, lower confidence limit of a two-sided CI with prescribed nominal level by conf_lv

ci_u

a scalar, upper confidence limit of a two-sided CI with prescribed nominal level by conf_lv

pval

p-value of Z-test, with null hypothesis that est is zero

Examples

trt <- list(est = 1.1, se = 0.2)
com <- list(est = 1.3, se = 0.18)
bucher(trt, com, conf_lv = 0.9)
#> $est
#> [1] -0.2
#> 
#> $se
#> [1] 0.2690725
#> 
#> $ci_l
#> [1] -0.6425848
#> 
#> $ci_u
#> [1] 0.2425848
#> 
#> $pval
#> [1] 0.4573036
#>