Centered patient data from two arm trial

Centered patient data from two arm trial

Usage

centered_ipd_twt

Format

A data frame with 1000 rows and 14 columns:

USUBJID

Unique subject identifiers for patients.

ARM

Assigned treatment arm.

AGE

Age in years at baseline.

SEX

Sex of patient recorded as character "Male"/"Female".

SMOKE

Smoking status at baseline as integer 1/0.

ECOG0

Indicator of ECOG score = 0 at baseline as integer 1/0.

N_PR_THER

Number of prior therapies received as integer 1, 2, 3, 4.

SEX_MALE

Indicator of SEX == "Male" as numeric 1/0.

AGE_CENTERED

Age in years at baseline relative to average in aggregate data agd.

AGE_MEDIAN_CENTERED

AGE greater/less than MEDIAN_AGE in agd coded as 1/0 and then centered at 0.5.

AGE_SQUARED_CENTERED

AGE squared and centered with respect to the AGE in agd. The squared age in the aggregate data is derived from the \(E(X^2)\) term in the variance formula.

SEX_MALE_CENTERED

SEX_MALE centered by the proportion of male patients in agd

ECOG0_CENTERED

ECOG0 centered by the proportion of ECOG0 in agd

SMOKE_CENTERED

SMOKE centered by the proportion of SMOKE in agd

N_PR_THER_MEDIAN_CENTERED

N_PR_THER centered by the median in agd.

See also