Check to see if weights are optimized correctly

This function checks to see if the optimization is done properly by checking the covariate averages before and after adjustment.

Usage

check_weights(weighted_data, processed_agd)

# S3 method for maicplus_check_weights
print(
  x,
  mean_digits = 2,
  prop_digits = 2,
  sd_digits = 3,
  digits = getOption("digits"),
  ...
)

Arguments

weighted_data

object returned after calculating weights using estimate_weights

processed_agd

a data frame, object returned after using process_agd or aggregated data following the same naming convention

x

object from check_weights

mean_digits

number of digits for rounding mean columns in the output

prop_digits

number of digits for rounding proportion columns in the output

sd_digits

number of digits for rounding mean columns in the output

digits

minimal number of significant digits, see print.default.

...

further arguments to print.data.frame

Value

data.frame of weighted and unweighted covariate averages of the IPD, and average of aggregate data

Methods (by generic)

print(maicplus_check_weights): Print method for check_weights objects

Examples

adsl <- read.csv(system.file("extdata", "adsl.csv",
  package = "maicplus",
  mustWork = TRUE
))
adrs <- read.csv(system.file("extdata", "adrs.csv",
  package = "maicplus",
  mustWork = TRUE
))
adtte <- read.csv(system.file("extdata", "adtte.csv",
  package = "maicplus",
  mustWork = TRUE
))

### AgD
# Baseline aggregate data for the comparator population
target_pop <- read.csv(system.file("extdata", "aggregate_data_example_1.csv",
  package = "maicplus", mustWork = TRUE
))
# target_pop2 <- read.csv(system.file("extdata", "aggregate_data_example_2.csv",
#                                     package = "maicplus", mustWork = TRUE))
# target_pop3 <- read.csv(system.file("extdata", "aggregate_data_example_3.csv",
#                                     package = "maicplus", mustWork = TRUE))

# for time-to-event endpoints, pseudo IPD from digitalized KM
pseudo_ipd <- read.csv(system.file("extdata", "psuedo_IPD.csv",
  package = "maicplus",
  mustWork = TRUE
))

#### prepare data ----------------------------------------------------------
target_pop <- process_agd(target_pop)
# target_pop2 <- process_agd(target_pop2) # demo of process_agd in different scenarios
# target_pop3 <- process_agd(target_pop3) # demo of process_agd in different scenarios
adsl <- dummize_ipd(adsl, dummize_cols = c("SEX"), dummize_ref_level = c("Female"))
use_adsl <- center_ipd(ipd = adsl, agd = target_pop)

match_res <- estimate_weights(
  data = use_adsl,
  centered_colnames = grep("_CENTERED$", names(use_adsl)),
  start_val = 0,
  method = "BFGS"
)
#> initial  value 500.000000 
#> iter  10 value 196.020742
#> iter  20 value 173.488262
#> final  value 173.420820 
#> converged

check <- check_weights(
  weighted_data = match_res,
  processed_agd = target_pop
)

print(check)
#>   covariate match_stat internal_trial internal_trial_after_weighted
#> 1       AGE       Mean         59.850                         51.00
#> 2       AGE     Median         59.000                         49.00
#> 3       AGE         SD          9.011                          3.25
#> 4  SEX_MALE       Prop          0.380                          0.49
#> 5     ECOG0       Prop          0.410                          0.35
#> 6     SMOKE       Prop          0.320                          0.19
#>   external_trial sum_centered_IPD_with_weights
#> 1          51.00                        0.0000
#> 2          49.00                        0.0000
#> 3           3.25                       -0.0041
#> 4           0.49                        0.0000
#> 5           0.35                        0.0000
#> 6           0.19                        0.0000