Generate time-to-event data

Function generates time-to-event data. We follow the data-generating mechanism from Bender et al. and Remiro-Azocar et al. 2021 to simulate Weibull-distributed survival times under a proportional hazards parameterization. As per Remiro-Azocar et al, we set the default scale parameter (lambda) to 8.5 and shape parameter (nu) to 1.3 to reflect frequently observed mortality trends in metastatic cancer patients. Censoring times are generated from exponential distribution where the rate parameter is 0.96 is selected to achieve a censoring rate of 35% under the active treatment at baseline (with the values of the covariates set to zero).

Usage

generate_survival_data(
  N = 150,
  ncov = 4,
  lambda = 8.5,
  nu = 1.3,
  censor_rate = 0.96,
  cor = 0.35,
  marginal_mean = 0.6,
  marginal_sd = 0.2,
  beta = rep(-log(0.67), ncov),
  gamma = sort(rep(c(0, -log(0.67)), length = ncov)),
  d0 = log(0.25),
  seed = 1
)

Arguments

N

sample size

ncov

number of covariates to generate

lambda

scale parameter in baseline hazard function

nu

shape parameter in baseline hazard function

censor_rate

rate parameter of the exponential distribution of censoring times

cor

correlation between covariates

marginal_mean

marginal mean of covariates

marginal_sd

marginal standard deviation of covariates

beta

vector of coefficients for prognostic factors. Defaults is -log(0.67) for all covariates.

gamma

vector of coefficients for effect modifiers. Default is 0 for half of the covariates and -log(0.67) for the others.

d0

coefficient for the treatment effect for patients with treat = 0

seed

seed used for simulation reproducibility

Value

data.frame with survival time, censor (event) status, treatment indicator, and covariates

References

Remiro-Azocar et al. 2021