Unanchored MAIC for time-to-event endpoint

Unanchored MAIC for time-to-event endpoint

Usage

maic_tte_unanchor(
  useWt,
  dat,
  dat_ext,
  trt,
  trt_ext,
  time_scale = "month",
  endpoint_name = "OS",
  transform = "log"
)

Arguments

useWt

a numeric vector of individual MAIC weights, length should the same as nrow(dat)

dat

a data frame that meet format requirements in 'Details', individual patient data (IPD) of internal trial

dat_ext

a data frame, pseudo IPD from digitized KM curve of external trial

trt

a character string, name of the interested treatment in internal trial (real IPD)

trt_ext

character string, name of the interested comparator in external trial used to subset dat_ext (pseudo IPD)

time_scale

a character string, 'year', 'month', 'week' or 'day', time unit of median survival time

endpoint_name

a character string, name of the endpoint

transform

a character string, pass to cox.zph

Value

A list of KM plot, analysis table, and diagnostic plot

Details

Format requirements for input dat and dat_ext are to have the following columns

treatment - character or factor column

status - logical column, TRUE for censored/death, FALSE for otherwise

time - numeric column, observation time of the status; unit in days