maicplus: Matching Adjusted Indirect Comparison

Facilitates performing matching adjusted indirect comparison (MAIC) analysis where the endpoint of interest is either time-to-event (e.g. overall survival) or binary (e.g. objective tumor response). The method is described by Signorovitch et al (2012) doi:10.1016/j.jval.2012.05.004 .

See also

Author

Maintainer: Isaac Gravestock isaac.gravestock@roche.com

Authors:

Gregory Chen xiangyi.gregory.chen@msd.com

Michael Seo michael.seo@roche.com

Other contributors:

Miranta Antoniou miranta.antoniou@roche.com [contributor]

Chrysostomos Kalyvas chrysostomos.kal@gmail.com [contributor]

MSD, Inc. [copyright holder, funder]

F. Hoffmann-La Roche AG [copyright holder, funder]